requirements.txt
fpdf2
pyautogui
pillow